fix: 기본 정보, 작가 프로필 수정 페이지 디테일 수정, 작가 api 경로 수정
